+++ glicol-rs/examples/plot_update.rs
