+++ glicol-rs/src/node/oscillator.rs
@@ -1,168 +1,3 @@
-use dasp_graph::{Buffer, Input, Node, NodeData, BoxedNodeSend};
-use pest::iterators::Pairs;
-use super::super::{Rule, EngineError, GlicolNodeData, NodeResult, mono_node};
-use super::{Para};
-
-/// Sine Wave Oscillator Node Builder
-pub struct SinOsc {
-    freq: f32,
-    phase: f32,
-    clock: usize,
-    buffer: Buffer<128>,
-    sidechain_info: Vec<u8>,
-    sr: usize,
-}
-
-impl SinOsc {
-    pub fn new() -> Self {
-        Self {
-            freq: 0.01,
-            phase: 0.,
-            clock: 0,
-            buffer: Buffer::<128>::default(),
-            sidechain_info: vec![],
-            sr: 44100,
-        }
-    }
-
-    pub fn freq(self, freq: f32) -> Self {
-        Self {freq, ..self}
-    }
-
-    pub fn sr(self, sr: usize) -> Self {
-        Self {sr, ..self}
-    }
-
-    pub fn build(self) -> GlicolNodeData {
-        mono_node! {
-            self
-        }
-    }
-}
-
-#[macro_export]
-macro_rules! sin {
-    {$($para: ident: $data:expr),*} => {
-         (
-            SinOsc::new()$(.$para($data))*.build()
-        )
-    }
-}
-
-// NodeData::new1(
-//     BoxedNodeSend::new(
-//         SinOsc {
-//             $($para: $data,)*
-//             ..Default::default()
-//         }
-//     )
-// )
-
-// impl SinOsc {
-//     pub fn new(paras: &str) -> GlicolNodeData {
-        // let mut sidechain_info = vec![];
-        // let paras: Vec<&str> = paras.split(" ").collect();
-        // let (freq, sr) = match paras.len() {
-        //     0 =>{ sidechain_info.push(0); (0.01, 44100) },
-        //     1 => { 
-        //         match paras[0].parse::<f32>() {
-        //             Ok(v) => (v, 44100),
-        //             Err(_) => (0.01, 44100)
-        //         }
-        //     },
-        //     2 => {
-        //         match paras[0].parse::<f32>() {
-        //             Ok(v) => (v, 44100),
-        //             Err(_) => (0.01, 44100)
-        //         }
-        //     },
-        //     _ => unimplemented!()
-        // };
-        // let freq = match freq.parse::<f32>() {
-        //     Ok(v) => v,
-        //     Err(_) => { sidechain_info.push(0); 0.01 }
-        // };
-        // let freq = match freq {
-        //     Para::Number(v) => v,
-        //     Para::Index(_) => { sidechain_info.push(0); 0.01 },
-        //     Para::Ref(_) => { sidechain_info.push(0); 0.01 },
-        //     _ => unimplemented!()
-        //     // Para::Ref(s) => { sidechain_info.push(s.to_string()); 0.01 },
-        //     // Para::Symbol(s) => unimplemented!()
-        // };
-    //     return mono_node!( Self {
-    //         freq,
-    //         phase: 0.,
-    //         clock: 0,
-    //         buffer: Buffer::<128>::default(),
-    //         sidechain_info,
-    //         sr,
-    //     })
-    // }
-    // handle_params!({
-    //     _freq: 440.0
-    // }, {
-    //     phase: 0.0,
-    //     clock: 0
-    // }, [(_freq, buffer, |_freq: f32|-> Buffer<128> {
-    //     Buffer::default()
-    // })]);
-// }
-
-/// The inputs.len() has two possible situation
-/// One is using Glicol as a standalone audio lib
-/// This will be zero, or any
-/// We should find out a way to differ standalone and live coding
-/// This is by seeing if the first input is a clock
-/// The clock is like 88280, 0, 0, 0
-impl Node<128> for SinOsc {
-    fn process(&mut self, inputs: &[Input<128>], output: &mut [Buffer<128>]) {
-        // if inputs.len() < 2 { return () };
-        if self.sidechain_info.len() == 1 {
-            if inputs.len() == 2 {
-                // has clock input
-                let clock = inputs[1].buffers()[0][0] as usize;
-
-                // avoid process twice
-                // without this, when use this node to control two different things
-                // the phase += will be called more than once and cause errors and mess
-                if self.clock != 0 && self.clock == clock {
-                    output[0] = self.buffer.clone();
-                    return ()
-                };
-                let mod_buf = &mut inputs[0].buffers();
-                for i in 0..128 {
-                    output[0][i] = (self.phase * 2.0 * std::f32::consts::PI).sin();
-                    self.phase += mod_buf[0][i] / self.sr as f32;
-                    if self.phase > 1.0 {
-                        self.phase -= 1.0
-                    }
-                }
-                self.buffer = output[0].clone();
-                self.clock = clock;
-            } else {
-                // in standalone mode, no mechanism to prevent double processing
-                let mod_buf = &mut inputs[0].buffers();
-                for i in 0..128 {
-                    output[0][i] = (self.phase * 2.0 * std::f32::consts::PI).sin();
-                    self.phase += mod_buf[0][i] / self.sr as f32;
-                    if self.phase > 1.0 {
-                        self.phase -= 1.0
-                    }
-                }
-            }
-        } else {
-            for i in 0..128 {
-                output[0][i] = (self.phase * 2.0 * std::f32::consts::PI).sin();
-                self.phase += self.freq / self.sr as f32;
-                if self.phase > 1.0 {
-                    self.phase -= 1.0
-                }
-            }
-        }
-    }
-}
-
 pub struct Impulse {
     clock: usize,
     period: usize,
